+++ Sources/DeclarationAccessibility/View/InformationView.swift
@@ -7,6 +7,7 @@
 // or see the "LICENSE" file for more details.
 
 #if os(iOS)
+import OUDSSwiftUI
 import SwiftUI
 
 // MARK: - Information View
@@ -15,13 +16,20 @@ import SwiftUI
 struct InformationView: View {
 
     var statement: Statement
-    var theme: Theme
+    var theme: OUDSTheme
+
+    @State private var showWebView = false
 
     var body: some View {
         ScrollView {
-            VStack(spacing: -20) {
-                GroupView(title: "date_title", subTitle: statement.auditDate)
-                    .accessibilityElement(children: .combine)
+            VStack(spacing: -10) {
+                if let auditDate = statement.auditDate {
+                    GroupView(title: "date_title", subTitle: auditDate.toDisplay)
+                        .accessibilityElement(children: .combine)
+                        .accessibilityLabel(auditDate.toVocalize.isEmpty
+                            ? auditDate.toDisplay
+                            : "\(NSLocalizedString("date_title", bundle: .module, comment: "")), \(auditDate.toVocalize)")
+                }
                 GroupView(title: "identity_title", subTitle: statement.identityName, text: statement.identityAddress)
                     .accessibilityElement(children: .combine)
                 GroupView(title: "referential_title", subTitle: statement.referentialName)
@@ -34,25 +42,25 @@ struct InformationView: View {
                 if statement.mustUseWebView {
                     NavigationLink(
                         destination: WebViewPage(url: statement.detailUrl),
-                        label: {
-                            Text("detail_button", bundle: .module)
-                                .padding()
-                                .background(theme.buttonColor)
-                                .foregroundColor(theme.foregroundColor)
-                                .font(.title3)
-                        })
+                        isActive: $showWebView)
+                    {
+                        EmptyView()
+                    }
+                    .hidden()
+
+                    OUDSButton(text: NSLocalizedString("detail_button", bundle: .module, comment: ""),
+                               appearance: .strong)
+                    {
+                        showWebView = true
+                    }
                 } else {
-                    Button(action: {
+                    OUDSButton(text: NSLocalizedString("detail_button", bundle: .module, comment: ""),
+                               appearance: .strong)
+                    {
                         if let url = URL(string: statement.detailUrl) {
                             UIApplication.shared.open(url)
                         }
-                    }, label: {
-                        Text("detail_button", bundle: .module)
-                            .padding()
-                            .background(theme.buttonColor)
-                            .foregroundColor(theme.foregroundColor)
-                            .font(.title3)
-                    })
+                    }
                 }
             }
         }
@@ -67,6 +75,8 @@ private struct GroupView: View {
     let subTitle: String
     let text: String?
 
+    @Environment(\.theme) private var theme
+
     init(title: LocalizedStringKey, subTitle: String, text: String? = nil) {
         self.title = title
         self.subTitle = subTitle
@@ -74,20 +84,19 @@ private struct GroupView: View {
     }
 
     var body: some View {
-        VStack(alignment: .leading, spacing: 5) {
+        VStack(alignment: .leading, spacing: theme.spaces.fixedXsmall) {
             Text(title, bundle: .module)
-                .font(.title3)
-                .bold()
+                .headingMedium(theme)
                 .multilineTextAlignment(.leading)
                 .frame(maxWidth: .infinity, alignment: .leading)
             Text(subTitle)
-                .font(.body)
+                .bodyDefaultLarge(theme)
             if let text {
                 Text(text)
-                    .font(.body)
+                    .bodyDefaultLarge(theme)
             }
         }
-        .padding()
+        .padding(theme.spaces.fixedSmall)
     }
 }
 
@@ -98,8 +107,11 @@ private struct WebViewPage: View {
 
     // swiftlint:disable force_unwrapping
     var body: some View {
-        StatementWebView(from: URL(string: url)!) // TODO: Manage this error case
-            .navigationTitle("Details") // TODO: Hard-coded not localized wording
+        StatementWebView(from: URL(string: url)!) // FIXME: Manage this error case
+            .navigationTitle(NSLocalizedString( // FIXME: Change API
+                "declaration_title",
+                bundle: .module,
+                comment: ""))
     }
     // swiftlint:enable force_unwrapping
 }
@@ -109,7 +121,7 @@ private struct WebViewPage: View {
 struct InformationView_Previews: PreviewProvider {
     static var previews: some View {
         InformationView(statement: Statement(
-            auditDate: "Test Audit Date",
+            auditDate: (toDisplay: "Test Audit Date", toVocalize: ""),
             referentialName: "Test Referential Name",
             referentialVersion: "Test Referential Version",
             referentialLevel: "Test Referential Level",
@@ -117,8 +129,9 @@ struct InformationView_Previews: PreviewProvider {
             detailUrl: "https://example.com",
             identityAddress: "Test Identity Address",
             identityName: "Test Identity Name"),
-        theme: .orange)
+        theme: WireframeTheme())
             .environment(\.locale, .init(identifier: "fr"))
+            .wireframePreview()
     }
 }