+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0](https://github.com/Orange-OpenSource/accessibility-statement-lib-ios/compare/2.0.0...2.1.0) - 2026-01-27
+
+### Added
+
+- Support of locales ca, es, nl, pl, ro, ru, sk (Orange-OpenSource/accessibility-statement-lib-ios#46)
+- Integration of OUDS library (version 1.1.0-rc.1) (Orange-OpenSource/accessibility-statement-lib-ios#38)
+
+### Changed
+
+- GitHub Actions workflow dedicated to snapshots
+- Add XCTestPlan for unit tests
+- Update README with Swift Package Index badge (Orange-OpenSource/accessibility-statement-lib-ios#47)
+- Generation of documentation and upload for GitHub Pages (Orange-OpenSource/accessibility-statement-lib-ios#50)
+
+### Fixed
+
+- When using VoiceOver, the audit date month was abbreviated and not vocalized correctly (Orange-OpenSource/accessibility-statement-lib-ios#54)
+- Hard-coded "Details" word instead of localized resource in webview title (Orange-OpenSource/accessibility-statement-lib-ios#53)
+
 ## [2.0.0](https://github.com/Orange-OpenSource/accessibility-statement-lib-ios/compare/1.3.0...2.0.0) - 2026-01-22
 
 ### Added